src/requirements.txt
flask
flask-sqlalchemy
flask-restful
Flask-HTTPAuth
psycopg2-binary
passlib
